[Decomp][ObjectMgr] Cohesion split: extract spell-data loaders to ObjectMgrSpellData.cpp
